Refactored to move into separate deeplinks-controller-api module
