+++ settings.gradle
@@ -50,6 +50,7 @@ include ':simplified-cardcreator'
 include ':simplified-content-api'
 include ':simplified-crashlytics'
 include ':simplified-crashlytics-api'
+include ':simplified-deeplinks-controller-api'
 include ':simplified-documents'
 include ':simplified-feeds-api'
 include ':simplified-files'
