--- simplified-accounts-registry/src/main/java/org/nypl/simplified/accounts/registry/ScreenID.kt
@@ -1,5 +1,0 @@
-package org.nypl.simplified.accounts.registry
-
-enum class ScreenID {
-  LOGIN, UNSPECIFIED
-}
