+++ simplified-books-controller/build.gradle
@@ -32,7 +32,8 @@ dependencies {
   implementation libs.palace.drm.core
   implementation libs.palace.http.api
   implementation libs.slf4j
+    implementation project(path: ':simplified-deeplinks-controller-api')
 
-  compileOnly libs.google.auto.value
+    compileOnly libs.google.auto.value
   annotationProcessor libs.google.auto.value.processor
 }
