+++ simplified-deeplinks-controller-api/README.md
@@ -1,0 +1,5 @@
+org.librarysimplified.deeplinks.controller.api
+===
+
+The `org.librarysimplified.deeplinks.controller.api` module specifies
+the _controller_ API for _deep links_ functionality.
