+++ simplified-deeplinks-controller-api/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=org.librarysimplified.deeplinks.controller.api
+POM_DESCRIPTION=Library Simplified (Deep Links controller API)
+POM_NAME=org.librarysimplified.deeplinks.controller.api
+POM_PACKAGING=aar
