+++ simplified-deeplinks-controller-api/src/main/java/org/nypl/simplified/deeplinks/controller/api/DeepLinkEvent.kt
@@ -1,4 +1,4 @@
-package org.nypl.simplified.accounts.registry
+package org.nypl.simplified.deeplinks.controller.api
 
 import org.nypl.simplified.accounts.api.AccountID
