+++ simplified-deeplinks-controller-api/src/main/java/org/nypl/simplified/deeplinks/controller/api/DeepLinksControllerType.kt
@@ -1,4 +1,4 @@
-package org.nypl.simplified.accounts.registry
+package org.nypl.simplified.deeplinks.controller.api
 
 import io.reactivex.Observable
 import org.nypl.simplified.accounts.api.AccountID
