+++ simplified-deeplinks-controller-api/src/main/java/org/nypl/simplified/deeplinks/controller/api/ScreenID.kt
@@ -1,0 +1,5 @@
+package org.nypl.simplified.deeplinks.controller.api
+
+enum class ScreenID {
+  LOGIN, UNSPECIFIED
+}
