+++ simplified-main/build.gradle
@@ -39,6 +39,7 @@ dependencies {
   api project(":simplified-cardcreator")
   api project(":simplified-content-api")
   api project(":simplified-crashlytics-api")
+  api project(":simplified-deeplinks-controller-api")
   api project(":simplified-documents")
   api project(":simplified-files")
   api project(":simplified-lcp")
